+++ test_planning/test_sawyer_planning.py
@@ -78,16 +78,15 @@ def main():
         # Use parametric linear interpolation with 10 steps between points.
         interp = partial(parametric_lerp, steps=10)
         # See params for PRM specific parameters
-        prm = PRM(state_space, svc, interp, params={'n_samples': 3000, 'k': 6, 'ball_radius': 2.0})
+        prm = PRM(state_space, svc, interp, params={'n_samples': 6000, 'k': 10, 'ball_radius': 1.70})
         logger.info("Planning....")
         plan = prm.plan(np.array([0, 0, 0, 0, 0, 0, 0]), np.array([1.5262755737449423, -0.1698540226273928, 2.7788151824762055, 2.4546623466066135, 0.7146948867821279, 2.7671787952787184, 2.606128412644311]))
         # get_path() reuses the interp function to get the path between vertices of a successful plan 
         path = prm.get_path(plan)
     if len(path) == 0:
-            logger.info("Planning failed....")
-            sys.exit(1)
+        logger.info("Planning failed....")
+        sys.exit(1)
     logger.info("Plan found....")
-    print(len(path))
     ##########
     # SPLINE #
     ##########
